🚚 Move: `Role` enum to `role` module [#63]
