+++ src/model/user.rs
@@ -2,7 +2,7 @@
 
 use crate::{
     client::{error::RspErr, Client},
-    model::cache::CacheData,
+    model::{cache::CacheData, role::Role},
     util::{deserialize_from_non_str_to_none, max_f64, to_unix_ts},
 };
 use serde::Deserialize;
@@ -478,119 +478,6 @@ impl User {
 impl AsRef<User> for User {
     fn as_ref(&self) -> &Self {
         self
-    }
-}
-
-/// The user's role.
-#[derive(Clone, Debug, Deserialize)]
-pub enum Role {
-    /// The normal user.
-    #[serde(rename = "user")]
-    User,
-    /// The anonymous user.
-    #[serde(rename = "anon")]
-    Anon,
-    /// The bot.
-    #[serde(rename = "bot")]
-    Bot,
-    /// The SYSOP.
-    #[serde(rename = "sysop")]
-    Sysop,
-    /// The administrator.
-    #[serde(rename = "admin")]
-    Admin,
-    /// The moderator.
-    #[serde(rename = "mod")]
-    Mod,
-    /// The community moderator.
-    #[serde(rename = "halfmod")]
-    Halfmod,
-    /// The banned user.
-    #[serde(rename = "banned")]
-    Banned,
-    /// The hidden user.
-    #[serde(rename = "hidden")]
-    Hidden,
-}
-
-impl Role {
-    /// Whether the user is an anonymous.
-    pub fn is_anon(&self) -> bool {
-        matches!(self, Role::Anon)
-    }
-
-    /// Whether the user is a bot.
-    pub fn is_bot(&self) -> bool {
-        matches!(self, Role::Bot)
-    }
-
-    /// Whether the user is a SYSOP.
-    pub fn is_sysop(&self) -> bool {
-        matches!(self, Role::Sysop)
-    }
-
-    /// Whether the user is an administrator.
-    pub fn is_admin(&self) -> bool {
-        matches!(self, Role::Admin)
-    }
-
-    /// Whether the user is a moderator.
-    pub fn is_mod(&self) -> bool {
-        matches!(self, Role::Mod)
-    }
-
-    /// Whether the user is a community moderator.
-    pub fn is_halfmod(&self) -> bool {
-        matches!(self, Role::Halfmod)
-    }
-
-    /// Whether the user is banned.
-    pub fn is_banned(&self) -> bool {
-        matches!(self, Role::Banned)
-    }
-
-    /// Whether the user is hidden.
-    pub fn is_hidden(&self) -> bool {
-        matches!(self, Role::Hidden)
-    }
-}
-
-impl AsRef<Role> for Role {
-    fn as_ref(&self) -> &Self {
-        self
-    }
-}
-
-impl ToString for Role {
-    /// Converts the given value to a `String`.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # use tetr_ch::model::user::Role;
-    /// assert_eq!(Role::User.to_string(), "User");
-    /// assert_eq!(Role::Anon.to_string(), "Anonymous");
-    /// assert_eq!(Role::Bot.to_string(), "Bot");
-    /// assert_eq!(Role::Sysop.to_string(), "SYSOP");
-    /// assert_eq!(Role::Admin.to_string(), "Administrator");
-    /// assert_eq!(Role::Mod.to_string(), "Moderator");
-    /// assert_eq!(Role::Halfmod.to_string(), "Community moderator");
-    /// assert_eq!(Role::Banned.to_string(), "Banned user");
-    /// assert_eq!(Role::Hidden.to_string(), "Hidden user");
-    /// ```
-    fn to_string(&self) -> String {
-        match self {
-            Role::User => "User",
-            Role::Anon => "Anonymous",
-            Role::Bot => "Bot",
-            Role::Sysop => "SYSOP",
-            Role::Admin => "Administrator",
-            Role::Mod => "Moderator",
-            Role::Halfmod => "Community moderator",
-            Role::Banned => "Banned user",
-            Role::Hidden => "Hidden user",
-        }
-        .to_string()
     }
 }
